+++ acme_windowing_appkit/acme_notification_callback.mm
@@ -11,7 +11,7 @@
 #include "acme/operating_system/argcargv.h"
 bool ns_get_dark_mode();
 
-void system_id_update(::platform::system * psystem, long long iUpdate, long long iPayload);
+void system_id_update(::platform::system * psystem, int iUpdate, long long iPayload);
 
 ::platform::system * application_system(::platform::application * papplication);
