+++ acme_windowing_appkit/macos_app.mm
@@ -26,7 +26,7 @@ void application_handle_command(::platform::application * papplication, const ch
 
 ::platform::system * application_system(::platform::application * papplication);
 
-void system_id_update(::platform::system * psystem, long long iUpdate, long long iPayload);
+void system_id_update(::platform::system * psystem, int iUpdate, long long iPayload);
 
 void node_will_finish_launching(::platform::system * psystem);
 void system_on_open_untitled_file(::platform::system * psystem);
